refactor: Move repair related logic to new view
